リファクタリング。
Database系が太ってきたので，

１。DatabaseUtilには接続関係と，共通で使う構文

２。それぞれのtableをいじるためのSQLは
「DBUホニャララ」

にクラス分割
