+++ src/main/java/tech/inudev/profundus/database/DBUMoney.java
@@ -1,0 +1,116 @@
+package tech.inudev.profundus.database;
+
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+
+public class DBUMoney extends DatabaseUtil {
+
+	final static Table table = Table.MONEY;
+	
+	static final String createStr = """
+            'name' VARCHAR(36) NOT NULL,
+            'amount' INT NOT NULL,
+            PRIMARY KEY ('name')
+			""";
+
+
+    /**
+     * Databaseに新たに金額データを登録する。
+     * 処理が完了した場合は、トランザクションをコミットする。
+     * 一方で、処理中にエラーが発生した場合は、トランザクションをロールバックする。
+     *
+     * @param name 金額データの名前
+     */
+    public static void createMoneyRecord(String name) {
+        try {
+            if (loadMoneyAmount(name) != null) {
+                throw new SQLException();
+            }
+
+            PreparedStatement preparedStatement = getConnection().prepareStatement("""
+                    INSERT INTO money (name, amount) VALUES (?, 0)
+                    """);
+            preparedStatement.setString(1, name);
+            preparedStatement.execute();
+            preparedStatement.close();
+
+            commitTransaction();
+        } catch (SQLException e) {
+            try {
+                getConnection().rollback();
+            } catch (SQLException e2) {
+                throw new RuntimeException(e2);
+            }
+        }
+    }
+
+    /**
+     * Databaseのnameに対応する金額データを取得する。
+     * 処理中にエラーが発生した場合は、トランザクションをロールバックする。
+     *
+     * @param name Databaseの検索に使用する金額データの名前
+     * @return 金額。Database上にデータが存在しなければnullを返す
+     */
+    public static Integer loadMoneyAmount(String name) {
+        try {
+
+            ResultSet resultSet = select(table,"name='"+name+"'");
+            Integer result = resultSet.next() ? resultSet.getInt("amount") : null;
+            resultSet.close();
+            return result;
+        } catch (SQLException e) {
+            try {
+                getConnection().rollback();
+                return null;
+            } catch (SQLException e2) {
+                throw new RuntimeException(e2);
+            }
+        }
+    }
+
+    /**
+     * 送金処理のトランザクションを実行する。
+     * 処理が完了した場合は、トランザクションをコミットする。
+     * 一方で、処理中にエラーが発生した場合は、トランザクションはロールバックする。
+     *
+     * @param selfName      自分の金額データの名前
+     * @param selfAmount    自分の金額データの金額
+     * @param partnerName   相手の金額データの名前
+     * @param partnerAmount 相手の金額データの金額
+     */
+    public static void remitTransaction(
+            String selfName,
+            int selfAmount,
+            String partnerName,
+            int partnerAmount) {
+        updateMoneyAmount(selfName, selfAmount);
+        updateMoneyAmount(partnerName, partnerAmount);
+        commitTransaction();
+    }
+
+    private static void updateMoneyAmount(String bankName, int amount) {
+        try {
+            if (loadMoneyAmount(bankName) == null) {
+                throw new SQLException();
+            }
+
+            PreparedStatement preparedStatement = getConnection().prepareStatement("""
+                    UPDATE money SET amount=? WHERE name=?
+                    """);
+            preparedStatement.setInt(1, amount);
+            preparedStatement.setString(2, bankName);
+            preparedStatement.execute();
+            preparedStatement.close();
+        } catch (SQLException e) {
+            try {
+            	getConnection().rollback();
+            } catch (SQLException e2) {
+                throw new RuntimeException(e2);
+            }
+        }
+    }
+
+
+}
